Sistema de filtros e distribuições aparentemente funcionando bem. Ainda trabalhando nele, nos detalhes.

+++ src/main/java/br/ifsul/enemsim/repositories/HabilidadeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ifsul.enemsim.entidades.Habilidade;
 
-public interface HabilidadeRepository extends JpaRepository<Habilidade, Integer> {
+public interface HabilidadeRepository extends JpaRepository<Habilidade, Byte> {
 
 }
